Refactored route and module name to 'vault' to reflect the encryption method.
