+++ lib/control/v1/index.js
@@ -7,7 +7,7 @@ module.exports = function Index(app) {
     res.render('index', {title: 'ACS'});
   });
 
-  app.post('/v1/encrypt', require('./encrypt')());
+  app.post('/v1/vault', require('./vault')());
 
   app.use(Err);
 };
